Move search to /search
This mirrors how we decided to do it in snapcraft.io, which was quite
carefully considered. We believe that "domain.com/search?q=" is a
more descriptive URL for the user to see than "domein.com/?q=". "/?q="
might be better for an API, but for a website, /search is more usable.

This also has added benefits for development, because "/search" leading
to the "search.html" template is more intuitive, and we can now keep the
logic for "/search" in its own routing function.