Stop checking cache for every request
Checking the cache for normal page requests wasn't working very well.
Page requests could be slow for any number of reasons - they always have
external dependencies that could be slow on connections. We'll have to
simply rely on checking the cache on the api.get method directly, which we already do.