Hardcode routes, remove local data
Replace:
- "/topics/<slug>" with explicit topic route ("/topics/maas" etc.)
- "/<group_slug>" with explicit group pages ("/cloud-and-server" etc.)

This is because each of these pages has hard-coded data. And since they
appear explicitly in the navigation, they are clearly intended to be
dedicated pages, rather than dynamic pages.

This structure is more declarative and much clearer to work with.

And it means we can remove all the local data.