Cache in memory rather than disk
We have been caching requests to a cache.sqlite file on disk. This means
that the cache will persist if the process restarts, but it is slower.

In Kubernetes, there's not really a whole lot of point persisting the
cache between different processes, because a single container runs a
single process, and usually if the process crashes the container is
recreated.

Caching in memory is about 0.05 seconds quicker (curl responses are around 0.20s rather than 0.25s
in local testing). This is not insignificant.

However, in local development, the disk cache is genuinely useful, as
the process will be restarted a lot, and we don't want to have to put up
with 10 second page loads all the time. For that reason I use
REQUEST_CACHE_BACKEND=sqlite in local development.