Remove flake8 and black from requirements.txt
Not only is it a shame to install dev dependencies for the
production image, I just realised that actually
Black will probably fail to install in Jenkins, as it requires
Python >= 3.6, and we only have 3.5 in Jenkins.

So I've made sure black is installed in the
canonicalwebteam/dev:1.6.3 image, so the linter should always work.